refactor(tracker): rename orchestration->run_orchestration, drop unused private re-exports
